+++ engine/core/windows/EngineWin.cpp
@@ -135,7 +135,7 @@ namespace ouzel::core::windows
         if (charCount == 0)
             throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to convert UTF-8 to wide char"};
 
-        auto buffer = std::make_unique<WCHAR[]>(charCount);
+        auto buffer = std::make_unique<WCHAR[]>(static_cast<std::size_t>(charCount));
         if (MultiByteToWideChar(CP_UTF8, 0, url.c_str(), -1, buffer.get(), charCount) == 0)
             throw std::system_error{static_cast<int>(GetLastError()), std::system_category(), "Failed to convert UTF-8 to wide char"};
